fixing ring attention issue, adding SequenceClassification Trainer (requested #169), fixing llama issue for llama3.1+, fixing mamba1/2 issues in float16

+++ src/easydel/modules/mamba2/modelling_mamba2_flax.py
@@ -536,7 +536,7 @@ class FlaxMamba2Mixer(nn.Module):
 				)
 				* (jnp.log(self.config.time_step_max) - jnp.log(self.config.time_step_min))
 				+ jnp.log(self.config.time_step_min)
-			),
+			).astype(jnp.float32),
 			1e9,
 		)
